requirements.txt
matplotlib
six
numpy>=1.9.0
pyopencl>=2013.1
scipy>=0.12.1
mot
pyyaml
nibabel
